Refactor voice assistant into focused components
- Replace ChatbotCTA with separate, focused components:
  - AIAssistant: Core AI functionality (TourPlayer + Dialog)
  - AICTABanner: Prominent CTA for page content flow
  - AskAIMini: Compact CTA for section embedding
- Fix TourPlayer bugs:
  - AnimatePresence exit animation now triggers correctly
  - Auto-advance timeout properly cleared on pause
  - Animation delays use inline styles instead of non-standard classes
- Remove floating action button (FAB) per user feedback
- Keep Speechify-style sidebar for tour interaction
- AIAssistant persists in RootLayout across all pages
