+++ src/components/layout/RootLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom"
 import { Header } from "@/components/shared/Header"
 import { Footer } from "@/components/shared/Footer"
-import { ChatbotCTA } from "@/components/voice-assistant"
+import { AIAssistant } from "@/components/voice-assistant"
 
 export function RootLayout() {
   return (
@@ -11,8 +11,8 @@ export function RootLayout() {
         <Outlet />
       </main>
       <Footer />
-      {/* AI Assistant - persists across all pages */}
-      <ChatbotCTA />
+      {/* AI Assistant - TourPlayer and Dialog, persists across all pages */}
+      <AIAssistant />
     </div>
   )
 }
