+++ src/components/voice-assistant/index.ts
@@ -1,4 +1,6 @@
 export { AskAboutMe } from "./AskAboutMe"
-export { ChatbotCTA, AskAIMini } from "./ChatbotCTA"
+export { AIAssistant } from "./AIAssistant"
+export { AICTABanner } from "./AICTABanner"
+export { AskAIMini } from "./AskAIMini"
 export { TourPlayer } from "./TourPlayer"
 export { danielContext, generateSystemPrompt, getFallbackResponse } from "./danielContext"
